requirements.txt
anyio==4.2.0
arrow==1.3.0
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
easygui==0.98.3
flet==0.19.0
flet-core==0.19.0
flet-runtime==0.19.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
Jinja2==3.1.3
libgravatar==1.0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
Pygments==2.17.2
pyhtml2pdf==0.0.7
PyMySQL==1.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
websocket-client==1.7.0
websockets==11.0.3
wsproto==1.2.0
infisical_api==1.0.0